+++ CHANGES.rst
@@ -36,6 +36,8 @@ Improvements
   :user:`SegiNyn`)
 - Clone registration tags when cloning registration forms and preserve registration
   tags when cloning registrations (:pr:`6964`)
+- Allow restricting reminder recipients by registration form and tags (:pr:`6877`,
+  thanks :user:`tomako, unconventionaldotdev`)
 
 Bugfixes
 ^^^^^^^^